refactor: renamed getTweets to getTemporaryTweets
